feat : move the server file outside nextjs app folder
